requirements.txt
streamlit>=1.33.0

langchain==0.3.26
langchain-openai==0.3.24
langchain-community==0.3.26
langchain-pinecone==0.2.8

sentence-transformers==4.1.0
pypdf==5.6.1

pinecone>=7.0.1

protobuf>=5.26.1
packaging>=24.2

python-dotenv==1.1.0
gspread
oauth2client